@@ -12,12 +12,12 @@ from drf_client.connection import Api as RestApi
 class BaseFacade:
     """Stores key static information used across the package."""
 
-    api: RestApi or None = None
-    api_options: dict or None = None
-    cmd_args: Namespace or None = None
+    api: RestApi | None = None
+    api_options: dict | None = None
+    cmd_args: Namespace | None = None
 
     @staticmethod
-    def initialize_api(api_options: dict, cmd_args: Namespace = None):
+    def initialize_api(api_options: dict, cmd_args: Namespace | None = None):
         """Initialize API with the given options."""
         if BaseFacade.api is None:
             # Only initialize ones
