@@ -2,15 +2,15 @@
 
 import unittest
 
+from .test_api import ApiTestCase
+from .test_resources import ResourceTestCase
+
 
 def get_tests():
     return full_suite()
 
 
 def full_suite():
-    from .api import ApiTestCase
-    from .resources import ResourceTestCase
-
     resourcesuite = unittest.TestLoader().loadTestsFromTestCase(ResourceTestCase)
     apisuite = unittest.TestLoader().loadTestsFromTestCase(ApiTestCase)
 
